refactor: Apply code review
- 단/복수가 정상적으로 적용되지 않은 폴더 수정
- 변수의 의미가 모호한 부분을 명확하게 수정
